Extension: Loop optimizations and class sorting
Optimized the loops by removing unecessary variables, and getting rid of the switch statment in favor of just refering to an array storing the letter days (allows the loop's value to dirrectly correspond to the letter days in the proper order). In addition, added a loop which goes through each row and distributes the classes by letter day.